Add HandlingFiles.py
Added the script HandlingFiles.py. 
In this script, we have 3 functions. 
The first one receives a CSV file as a parameter and returns a list of dictionaries from that file. The second one receives the list of dictionaries, employee_list as a parameter and return a dictionary of department : amount. And the last one writes a dictionary of department : amount to a txt file.