Add Health_Check.py
Added the script Health_Check.py. 
In this script, we have a Health Check for our computer. 
We have two functions. One checks the disk usage and the other one checks the cpu usage.
Finally, if the disk usage in root directory is different from 0 (which is true if there is disk space) and if the CPU usage is less than 75% it prints the message "Everything is OK!". But if there is no space left, or if the disk usage is more than 75% it prints "ERROR"