Add 10_supplier_image_upload.py
This Python script is used to upload a set of images to a specified URL. Here’s what it does:

It imports the necessary modules: requests for making HTTP requests and os for interacting with the operating system.
It sets the directory variable to the path where the images are stored.
It sets the url variable to the URL where the images will be uploaded. Note that this is a placeholder URL and should be replaced later.
It then loops over all the files in the specified directory.
For each file that ends with the extension “.jpeg”, it opens the image file in binary mode.
It then makes a POST request to the specified URL, uploading the opened file. The ‘file’ parameter in the files dictionary is where the opened file data is passed.

In summary, this script uploads all “.jpeg” images in a specified directory to a specified URL.